requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
openai>=1.30.0
azure-identity==1.18.0
